Start making .c files self-contained (despite being a unity build) so that go-to-definition and LSP stuff works
