+++ src/main.c
@@ -40,6 +40,7 @@
 #include "cstring.c"
 #include "path_utils.c"
 #include "os_utils.c"
+#include "print_floats.c"
 #include "print.c"
 #include "array.c"
 #include "hash_map.c"
@@ -47,6 +48,7 @@
 #include "linker.c"
 #include "lexer/module.c"
 #include "ast/module.c"
+#include "ast/print_ast.c"
 #include "parser/module.c"
 #include "resolver/module.c"
 #include "bytecode/module.c"
